requirements.txt
asgiref==3.7.2
azure-core==1.30.1
azure-storage-blob==12.19.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-tailwind==1.0.3
cryptography==42.0.5
Django==5.0.2
django-appconf==1.0.6
django-compressor==4.4
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.6
isodate==0.6.1
joblib==1.3.2
msal==1.27.0
msal-extensions==1.1.0
npm==0.1.1
npx==0.1.2
numpy==1.26.4
optional-django==0.1.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
portalocker==2.8.2
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-env==1.0.0
pytz==2024.1
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
scikit-learn==1.3.0
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
